\name{big_letters}
\alias{big_letters}

\title{
	Letters built
}
\description{
	The 26 or more lower-case letters of the Roman alphabet;
}

\usage{
	big_letters(len)
}

\arguments{
	  \item{len}{ Get or set the length of letters. }
}

\details{
}

\value{
}

\references{
}

\author{ Jae-seong Yoo <praster1@gmail.com> }

\note{
}

\seealso{
}

\examples{
}

\keyword{ }
