\name{BN_Data_Generator-package}
\alias{BN_Data_Generator-package}
\alias{BN_Data_Generator}
\docType{package}

\title{
	Data Generator based on Bayesian Network Model
}

\details{

	\tabular{ll}{
		Package: \tab BN_Data_Generator\cr
		Type: \tab Package\cr
		Version: \tab 1.0\cr
		Date: \tab 2014-12-24\cr
		License: \tab GPL (>=2)\cr
		}
}

\author{
	Jae-seong Yoo <praster1@gmail.com>
}

\references{
Jae-seong Yoo, (2014), "A Study on Comparison of Bayesian Network Structure Learning Algorithms for Selecting Appropriate Models", M.S. thesis, Department of Statistics, Korea University, Seoul.
}

\keyword{ Bayesian Network }

\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}

\examples{
~~ simple examples of the most important functions ~~
}
