\name{BN_Data_Generator}
\alias{BN_Data_Generator}

\title{
	Data Generator based on Bayesian Network Model
}

\description{
	Data Generator based on Bayesian Network Model
}

\usage{
	BN_Data_Generator(arcs_mat, input_Probs, n, node_names = NULL, cardinalities = NULL)
}

\arguments{
  \item{arcs_mat}{ A matrix that determines the arcs.}
  \item{input_Probs}{ The conditional probabilities. }
  \item{n}{ Data size. }
  \item{node_names}{ The names of each nodes. }
  \item{cardinalities}{ The cardinalities of each nodes. }
}

\details{
}

\value{
}

\references{
	Jae-seong Yoo, (2014), "A Study on Comparison of Bayesian Network Structure Learning Algorithms for Selecting Appropriate Models", M.S. thesis, Department of Statistics, Korea University, Seoul.
}

\author{ Jae-seong Yoo <praster1@gmail.com> }

\note{
}

\seealso{
}

\examples{
}

\keyword{
}
