\name{fromto_to_mat}
\alias{fromto_to_mat}

\title{
	Convert from 'fromto' to 'matrix'
}

\description{
	Convert from 'fromto' to 'matrix' that determines the arcs.
}

\usage{
	fromto_to_mat(fromto, node_names)
}

\arguments{
  \item{fromto}{ A matrix form structured 'fromto' that determines the arcs.}
  \item{node_names}{ The names of each nodes. }
}

\details{
}

\value{
}

\references{
}

\author{ Jae-seong Yoo <praster1@gmail.com> }

\note{
}

\seealso{
}

\examples{
}

\keyword{ }
